# apps/docs/openapi/v1/openapi.yaml
openapi: 3.0.3
info:
  title: AI Stats Gateway API
  description: >-
    A gateway API for accessing various AI models with OpenAI-compatible
    endpoints.
  version: 1.0.0
  contact:
    name: AI Stats
    url: https://docs.ai-stats.phaseo.app
    email: danielbutler500@gmail.com
servers:
  - url: https://api.phaseo.app/v1
security:
  - BearerAuth: []
paths:
...
  /webhook-endpoints:
    get:
      x-beta: true
      tags:
        - Platform
      operationId: listWebhookEndpoints
      summary: List webhook endpoints
      description: Lists workspace-managed async webhook endpoints.
      responses:
        '200':
          description: Webhook endpoint list
          content:
            application/json:
              schema:
                type: object
                properties:
                  object:
                    type: string
                    example: list
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/WebhookEndpoint'
    post:
      x-beta: true
      tags:
        - Platform
      operationId: createWebhookEndpoint
      summary: Create webhook endpoint
      description: >-
        Creates a workspace-managed async webhook endpoint and returns the
        signing secret once.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookEndpointCreateRequest'
      responses:
        '201':
          description: Created webhook endpoint
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookEndpointCreateResponse'
  /webhook-endpoints/{endpoint_id}:
    get:
      x-beta: true
      tags:
        - Platform
      operationId: getWebhookEndpoint
      summary: Get webhook endpoint
      parameters:
        - name: endpoint_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Webhook endpoint
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookEndpoint'
    patch:
      x-beta: true
      tags:
        - Platform
      operationId: updateWebhookEndpoint
      summary: Update webhook endpoint
      parameters:
        - name: endpoint_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookEndpointUpdateRequest'
      responses:
        '200':
          description: Updated webhook endpoint
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookEndpoint'
    delete:
      x-beta: true
      tags:
        - Platform
      operationId: deleteWebhookEndpoint
      summary: Delete webhook endpoint
      parameters:
        - name: endpoint_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Delete response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /webhook-endpoints/{endpoint_id}/rotate-secret:
    post:
      x-beta: true
      tags:
        - Platform
      operationId: rotateWebhookEndpointSecret
      summary: Rotate webhook endpoint secret
      parameters:
        - name: endpoint_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Webhook endpoint with new signing secret
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookEndpointCreateResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: Bearer token authentication
  schemas:
...
    WebhookEndpoint:
      type: object
      properties:
        id:
          type: string
        workspaceId:
          type: string
        name:
          type: string
        url:
          type: string
          format: uri
        status:
          type: string
          enum:
            - active
            - disabled
            - deleted
        events:
          type: array
          items:
            type: string
        hasSecret:
          type: boolean
        createdBy:
          type: string
          nullable: true
        createdAt:
          type: string
          nullable: true
        updatedAt:
          type: string
          nullable: true
        deletedAt:
          type: string
          nullable: true
    WebhookEndpointCreateRequest:
      type: object
      required:
        - url
      properties:
        name:
          type: string
        url:
          type: string
          format: uri
        events:
          type: array
          items:
            type: string
    WebhookEndpointUpdateRequest:
      type: object
      properties:
        name:
          type: string
        url:
          type: string
          format: uri
        events:
          type: array
          items:
            type: string
        status:
          type: string
          enum:
            - active
            - disabled
    WebhookEndpointCreateResponse:
      allOf:
        - $ref: '#/components/schemas/WebhookEndpoint'
        - type: object
          properties:
            signing_secret:
              type: string
              description: Returned once when the endpoint is created or rotated.
...